critical bug fix

+++ src/repositories/participants_repository.py
@@ -1,87 +1,42 @@
 import ormar
-from db import Participant
+import ormar.exceptions
+import db
 
 
 class ParticipantRepository:
-    # async def create(self, data: dict[str, any]):
-    #        res = await Participant.objects.create(**data)
-    #        return res
-    
-    # async def get(self, id):
-    #     try:
-    #         res = await Participant.objects.filter(giveaway_id=id).all()
-    #         return res
-    #     except ormar.NoMatch:
-    #         return None
-    
-    # async def get_all(self):
-    #     try:
-    #         res = await Participant.objects.all()
-    #         return res
-    #     except ormar.NoMatch:
-    #         return None
-
-    # async def get_by_id(self, id):
-    #     try:
-    #         res = await Participant.objects.filter(giveaway_id=id).all()
-    #         return res
-    #     except ormar.NoMatch:
-    #         return None
-    
-    # async def get_by_user_id(self, user_id):
-    #     try:
-    #         res = await Participant.objects.filter(user_id=user_id).all()
-    #         return res
-    #     except ormar.NoMatch:
-    #         return None
-    
-    # async def get_by_id_and_user_id(self, id, user_id):
-    #     try:
-    #         res = await Participant.objects.filter(giveaway_id=id, user_id=user_id).all()
-    #         return res
-    #     except ormar.NoMatch:
-    #         return None
-
-    # async def update(self, id, data):
-    #     res = await Participant.objects.update(giveaway_id=id, **data)
-    #     return res
-    
-    # async def delete(self, id):
-    #     res = await Participant.objects.filter(giveaway_id=id).delete()
-    #     return res
-    
-    # async def delete_by_ids(self, id, user_id):
-    #     res = await Participant.objects.filter(giveaway_id=id, user_id=user_id).delete()
-    #     return res
-
-
     async def create(self, data: dict[str, any]):
         try:
-            res = await Participant.objects.create(**data)
+            res = await db.Participant.objects.create(**data)
             return res
-        except ormar.ModelError:
+        except ormar.exceptions.ModelError:
             return None
 
     async def get(self, id=None, user_id=None):
         try:
-            if id:
-                return await Participant.objects.filter(giveaway_id=id).all()
-            elif user_id:
-                return await Participant.objects.filter(user_id=user_id).all()
+            if id is not None and user_id is not None:
+                to_return = await db.Participant.objects.get(giveaway_id=id, user_id=user_id)
+                print(to_return)
+                if not to_return:
+                    return None
+                return to_return
+            elif id is not None:
+                return await db.Participant.objects.filter(giveaway_id=id).all()
+            elif user_id is not None:
+                return await db.Participant.objects.filter(user_id=user_id).all()
             else:
-                return await Participant.objects.all()
+                return await db.Participant.objects.all()
         except ormar.NoMatch:
             return None
 
     async def delete(self, id=None, user_id=None):
         try:
-            if id and user_id:
-                return await Participant.objects.filter(giveaway_id=id, user_id=user_id).delete()
-            elif id:
-                return await Participant.objects.filter(giveaway_id=id).delete()
-            elif user_id:
-                return await Participant.objects.filter(user_id=user_id).delete()
+            if id is not None and user_id is not None:
+                return await db.Participant.objects.filter(giveaway_id=id, user_id=user_id).delete()
+            elif id is not None:
+                return await db.Participant.objects.filter(giveaway_id=id).delete()
+            elif user_id is not None:
+                return await db.Participant.objects.filter(user_id=user_id).delete()
             else:
                 return None
-        except ormar.ModelError:
+        except ormar.exceptions.ModelError:
             return None
